Refactor API and Stripe code
